pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dans-dv-upload"
version = "0.1.0"
dependencies = [
    "requests",
    "pyyaml",
]

[project.scripts]
dv-upload = "dans_dv_upload.dv_upload:main"

[tool.setuptools.package-dir]
"" = "src/main/python"

[project.optional-dependencies]
# tkinter is required for UI mode, but must be installed via your OS package manager (e.g., 'sudo apt install python3-tk' on Debian/Ubuntu/Mint)
tkinter = []
